Move Builder, Runner from kas_core to kas
