+++ crates/kas-core/src/event/cx/cx_pub.rs
@@ -722,8 +722,8 @@ impl<'a> EventCx<'a> {
     /// available to a running UI. This method may be used instead.
     ///
     /// Requirement: the type `Data` must match the type of data passed to the
-    /// [`Runner`](crate::runner::Runner) and used by other windows.
-    /// If not, a run-time error will result.
+    /// [`Runner`](https://docs.rs/kas/latest/kas/runner/struct.Runner.html)
+    /// and used by other windows. If not, a run-time error will result.
     ///
     /// Caveat: if an error occurs opening the new window it will not be
     /// reported (except via log messages).
