+++ crates/kas-core/src/theme/mod.rs
@@ -8,7 +8,7 @@
 //! Widgets expect the theme to provide an implementation of [`SizeCx`] and of
 //! [`DrawCx`].
 //!
-//! Constructing a [`Runner`](crate::runner::Runner) requires a [`Theme`]. Two
+//! Launching a UI requires a [`Theme`]. Two
 //! implementations are provided here: [`SimpleTheme`] and [`FlatTheme`].
 //! An adapter, [`MultiTheme`], is also provided.
